[MISC][RL] Rename file to kebab case
